Initial Conditions for Stochastic Individual Contact Models

Sets the initial conditions for stochastic individual contact models simulated with icm.

init_seiqhrf(
  s.num = 9997,
  e.num = 0,
  i.num = 3,
  q.num = 0,
  h.num = 0,
  r.num = 0,
  f.num = 0,
  ...
)

Arguments

s.num
Initial number of *S compartment individuals in the simulated population. An overall population of 10,000 is a good compromise. A set of models will still take several minutes or more to run, in parallel.

e.num
Initial number of E compartment individuals in the simulated population.

i.num
Initial number of I compartment individuals in the simulated population.

q.num
Initial number of Q compartment individuals in the simulated population.

h.num
Initial number of H compartment individuals in the simulated population.

r.num
Initial number of R compartment individuals in the simulated population.

f.num
Initial number of F compartment individuals in the simulated population.

...
Additional initial conditions passed to model.

Details

The initial conditions for a model solved with icm should be input into the init.icm function. This function handles initial conditions for both base models and original models using new modules. For an overview of initial conditions for base ICM class models, consult the Basic ICMs tutorial.

See also