Epidemic Parameters for Stochastic Individual Contact Models

Sets the epidemic parameters for stochastic individual contact models simulated with seiqhrf.

param_seiqhrf(
  inf.prob.e = 0.02,
  act.rate.e = 10,
  inf.prob.i = 0.05,
  act.rate.i = 10,
  inf.prob.q = 0.02,
  act.rate.q = 2.5,
  prog.rate = 1/10,
  quar.rate = 1/30,
  hosp.rate = 1/100,
  disch.rate = 1/15,
  rec.rate = 0.071,
  arec.rate = 0.05,
  prog.dist.scale = 5,
  prog.dist.shape = 1.5,
  quar.dist.scale = 1,
  quar.dist.shape = 1,
  hosp.dist.scale = 1,
  hosp.dist.shape = 1,
  disch.dist.scale = 1,
  disch.dist.shape = 1,
  rec.dist.scale = 35,
  rec.dist.shape = 1.5,
  arec.dist.scale = 35,
  arec.dist.shape = 1.5,
  fat.rate.base = 1/50,
  hosp.cap = 40,
  fat.rate.overcap = 1/25,
  fat.tcoeff = 0.5,
  vital = TRUE,
  a.rate = (10.5/365)/1000,
  a.prop.e = 0.01,
  a.prop.i = 0.001,
  a.prop.q = 0.01,
  ds.rate = (7/365)/1000,
  de.rate = (7/365)/1000,
  di.rate = (7/365)/1000,
  dq.rate = (7/365)/1000,
  dh.rate = (20/365)/1000,
  dr.rate = (7/365)/1000,
  ...
)

Arguments

inf.prob.e
Probability of passing on infection at each exposure event for interactions between infectious people in the E compartment and susceptible in S. Note the default is lower than for inf.prob.i reflecting the reduced infectivity of infected but asymptomatic people (the E compartment). Otherwise as for inf.exp.i.

act.rate.e
The number of exposure events (acts) between infectious individuals in the E compartment and susceptible individuals in the S compartment, per day. Otherwise as for act.rate.i.

inf.prob.i
Probability of passing on infection at each exposure event for interactions between infectious people in the I compartment and susceptible in S. Reducing inf.prob.i is equivalent to increasing hygiene measures, such as not putting hands in eyes, nose or moth, use of hand sanitisers, wearing masks by the infected, and so on.

act.rate.i
The number of exposure events (acts) between infectious individuals in the I compartment and susceptible individuals in the S compartment, per day. It's stochastic, so the rate is an average, some individuals may have more or less. Note that not every exposure event results in infection - that is governed by the inf.prob.i parameters (see below). Reducing act.rate.i is equivalent to increasing social distancing by people in the I compartment.

inf.prob.q
Probability of passing on infection at each exposure event for interactions between infectious people in the Q compartment and susceptible in S. Note the default is lower than for inf.prob.i reflecting the greater care that self-isolated individuals will, on average, take regarding hygiene measures, such as wearing masks, to limit spread to others. Otherwise as for inf.exp.i.

act.rate.q
The number of exposure events (acts) between infectious individuals in the Q compartment (isolated, self or otherwise) and susceptible individuals in the S compartment, per day. Note the much lower rate than for the I and E compartments, reflecting the much greater degree of social isolation for someone in (self-)isolation. The exposure event rate is not zero for this group, just much less. Otherwise as for act.rate.i.

prog.rate
Rate per day at which people who are infected but asymptomatic (E compartment) progress to becoming symptomatic (or test-positive), the I compartment. See prog.rand above for more details.

quar.rate
Rate per day at which symptomatic (or tested positive), infected I compartment people enter self-isolation (Q compartment). Asymptomatic E compartment people can't enter self-isolation because they don't yet know they are infected. Default is a low rate reflecting low community awareness or compliance with self-isolation requirements or practices, but this can be tweaked when exploring scenarios.

hosp.rate
Rate per day at which symptomatic (or tested positive), infected I compartment people or self-isolated Q compartment people enter the state of requiring hospital care -- that is, become serious cases. A default rate of 1 duration of about 10 days means a bit less than 10 require hospitalisation, which seems about right (but can be tweaked, of course).

disch.rate
Rate per day at which people needing hospitalisation recover.

rec.rate
Rate per day at which people who are infected and symptomatic (I compartment) recover, thus entering the R compartment. See rec.rand above for more details.

arec.rate
Rate per day at which people who are exposed but asymptotic (E compartment) recover, thus entering the R compartment. See arec.rand above for more details.

prog.dist.scale
Scale parameter for Weibull distribution for progression, see prog.rand for details.

prog.dist.shape
Shape parameter for Weibull distribution for progression, see prog.rand for details. Read up on the Weibull distribution before changing the default.

quar.dist.scale
Scale parameter for Weibull distribution for recovery, see quar.rand for details.

quar.dist.shape
Shape parameter for Weibull distribution for recovery, see quar.rand for details. Read up on the Weibull distribution before changing the default.

hosp.dist.scale
Scale parameter for Weibull distribution for recovery, see hosp.rand for details.

hosp.dist.shape
Shape parameter for Weibull distribution for recovery, see hosp.rand for details. Read up on the Weibull distribution before changing the default.

disch.dist.scale
Scale parameter for Weibull distribution for recovery, see disch.rand for details.

disch.dist.shape
Shape parameter for Weibull distribution for recovery, see disch.rand for details. Read up on the Weibull distribution before changing the default.

rec.dist.scale
Scale parameter for Weibull distribution for recovery, see rec.rand for details.

rec.dist.shape
Shape parameter for Weibull distribution for recovery, see rec.rand for details. Read up on the Weibull distribution before changing the default.

arec.dist.scale
Scale parameter for Weibull distribution for recovery, see arec.rand for details.

arec.dist.shape
Shape parameter for Weibull distribution for recovery, see arec.rand for details.

fat.rate.base
Baseline mortality rate per day for people needing hospitalisation (deaths due to the virus). See fat.rand for more details.

hosp.cap
Number of available hospital beds for the modelled population. See fat.rand for more details.

fat.rate.overcap
Mortality rate per day for people needing hospitalisation but who can't get into hospital due to the hospitals being full (see hosp.cap and fat.rand). The default rate is twice that for those who do get into hospital.

fat.tcoeff
Time co-efficient for increasing mortality rate as time in the H compartment increases for each individual in it. See fat.rand for details.

vital
Enables demographics, that is, arrivals and departures, to and from the simulated population.

a.rate
Background demographic arrival rate. Currently all arrivals go into the S compartment, the default is approximately the daily birth rate for Australia. Will be extended to cover immigration in future versions.

a.prop.e
Arrivals proportion that goes to E (immigration).

a.prop.i
Arrivals proportion that goes to I (immigration).

a.prop.q
Arrivals proportion that goes to Q (immigration).

ds.rate
Background demographic departure (death not due to virus) rates. Defaults based on Australian crude death rates. Can be used to model emigration as well as deaths.

de.rate
Background demographic departure (death not due to virus) rates. Defaults based on Australian crude death rates. Can be used to model emigration as well as deaths.

di.rate
Background demographic departure (death not due to virus) rates. Defaults based on Australian crude death rates. Can be used to model emigration as well as deaths.

dq.rate
Background demographic departure (death not due to virus) rates. Defaults based on Australian crude death rates. Can be used to model emigration as well as deaths.

dh.rate
Background demographic departure (death not due to virus) rates. Defaults based on Australian crude death rates. Can be used to model emigration as well as deaths.

dr.rate
Background demographic departure (death not due to virus) rates. Defaults based on Australian crude death rates. Can be used to model emigration as well as deaths.

...
Other parameters.

Details

param_seiqhrf sets the epidemic parameters for the stochastic individual contact models simulated with the seiqhrf function. Models may use the base types, for which these parameters are used, or new process modules which may use these parameters (but not necessarily). A detailed description of ICM parameterization for base models is found in the Basic ICMs tutorial.

For base models, the model specification will be chosen as a result of the model parameters entered here and the control settings in control_seiqhrf.

New Modules

To build original models outside of the base models, new process modules may be constructed to replace the existing modules or to supplement the existing set. These are passed into the control settings in control_seiqhrf. New modules may use either the existing model parameters named here, an original set of parameters, or a combination of both. The ... allows the user to pass an arbitrary set of original model parameters into param_seiqhrf. Whereas there are strict checks with default modules for parameter validity, these checks are the user's responsibility with new modules.

See also