Plot simulation result

Plot simulation result

# S3 method for seiqhrf
plot(
  x,
  method = NULL,
  comp_remove = "none",
  time_limit = 90,
  ci = TRUE,
  sep_compartments = FALSE,
  trans = FALSE,
  known = NULL,
  start_date = ymd("2020-03-21"),
  show_start_date = FALSE,
  x_axis = "Date (MM-DD)",
  plot_title = "",
  return_df = TRUE,
  market.share = 0.04,
  icu_percent = 0.1,
  sim_population = 1000,
  total_population = NULL,
  ...
)

Arguments

x
An seiqhrf object returned from function seiqhrf.

method
If "times", plot Duration frequency distributions. If "weekly_local", plot local weekly estimates from simulation. If NULL, plot sirplus plots.

comp_remove
Compartments to remove. Suggest c(s.num, r.num)

time_limit
Number of steps (days) to plot.

ci
T/F to include 95% confidence intervals in sirplus plot.

sep_compartments
T/F use faceting to show each compartment in a separate plot, only works if plotting a single simulation.

trans
Y-axis transformation (e.g. log2, log10).

known
Dataframe with known compartment numbers to plot alongside projections

start_date
Date for day 0.

show_start_date
First date to show in plots. Use ymd format. If FALSE, shows from step 1.

x_axis
Title for x-axis.

plot_title
Title for whole plot.

return_df
In effect only when method == "weekly", if TRUE returns also the dataframe used for plotting as well as the ggplot object.

market.share
between 0 and 1, percentage of local hospital beds in the simulated unit (e.g. state)

icu_percent
between 0 and 1, percentage of patients that should go to ICU among the ones that need hospitalization

sim_population
Size of population simulated. Only needed if providing `total_population`.

total_population
True population size, needed only if simulation size is smaller than the true population size due to computational cost etc.

...
Additional parameters