Extract Model Data for Stochastic Models

This function extracts model simulations for objects of classes seiqhrf into a data frame using the generic as.data.frame function.

# S3 method for seiqhrf
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  out = "vals",
  sim,
  qval,
  ...
)

Arguments

x
An EpiModel object of class icm or netsim.

row.names
See as.data.frame.default.

optional
See as.data.frame.default.

out
Data output to data frame: "mean" for row means across simulations, "sd" for row standard deviations across simulations, "qnt" for row quantiles at the level specified in qval, or "vals" for values from individual simulations.

sim
If out="vals", the simulation number to output. If not specified, then data from all simulations will be output.

qval
Quantile value required when out="qnt".

...
See as.data.frame.default.

Details

These methods work for both seiqhrf class models. The available output includes time-specific means, standard deviations, quantiles, and simulation values (compartment and flow sizes) from these stochastic model classes. Means, standard deviations, and quantiles are calculated by taking the row summary (i.e., each row of data is corresponds to a time step) across all simulations in the model output.

Examples

## Stochastic SEIQHRF model
param <- param_seiqhrf()
init <- init_seiqhrf(s.num = 500, i.num = 1)
control <- control_seiqhrf(nsteps = 10, nsims = 3)
mod <- seiqhrf(init, control, param)

# Default output all simulation runs, default to all in stacked data.frame
as.data.frame(mod)

#>    sim time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1    1    1   500     1 501       0       0       0         0       0       0
#> 2    1    2   498     1 501       1       0       0         0       0       0
#> 3    1    3   497     1 501       0       1       0         0       0       0
#> 4    1    4   496     1 501       0       0       0         0       0       0
#> 5    1    5   494     2 501       2       0       0         0       0       0
#> 6    1    6   491     3 501       2       1       0         0       0       0
#> 7    1    7   490     4 501       1       0       0         0       0       0
#> 8    1    8   486     4 501       2       0       1         0       0       0
#> 9    1    9   484     6 501       2       0       0         0       0       0
#> 10   1   10   483     8 501       1       1       0         0       0       0
#> 11   2    1   500     1 501       0       0       0         0       0       0
#> 12   2    2   499     1 501       1       0       0         0       0       0
#> 13   2    3   499     1 501       0       0       0         0       0       0
#> 14   2    4   496     1 501       1       0       0         0       0       0
#> 15   2    5   495     2 501       0       0       0         0       0       0
#> 16   2    6   494     3 501       1       1       0         0       0       0
#> 17   2    7   492     4 501       1       0       0         0       0       0
#> 18   2    8   485     4 501       4       0       0         1       0       0
#> 19   2    9   483     4 501       2       1       1         0       0       0
#> 20   2   10   478     7 501       2       0       0         0       0       0
#> 21   3    1   500     1 501       0       0       0         0       0       0
#> 22   3    2   500     1 501       0       0       0         0       0       0
#> 23   3    3   500     0 501       0       0       1         0       0       0
#> 24   3    4   500     0 501       0       0       0         0       0       0
#> 25   3    5   500     0 501       0       0       0         0       0       0
#> 26   3    6   500     0 501       0       0       0         0       0       0
#> 27   3    7   500     0 501       0       0       0         0       0       0
#> 28   3    8   500     0 501       0       0       0         0       0       0
#> 29   3    9   500     0 501       0       0       0         0       0       0
#> 30   3   10   500     0 501       0       0       0         0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#> 11       0       0       0       0       0      0        0        0        0
#> 12       0       0       0       0       0      0        0        0        0
#> 13       0       0       0       0       0      0        0        0        0
#> 14       0       0       0       0       0      0        0        0        0
#> 15       0       0       0       0       0      0        0        0        0
#> 16       0       0       0       0       0      0        0        0        0
#> 17       0       0       0       0       0      0        0        0        0
#> 18       0       0       0       0       0      0        0        0        0
#> 19       0       0       0       0       0      0        0        0        0
#> 20       0       0       0       0       0      0        0        0        0
#> 21       0       0       0       0       0      0        0        0        0
#> 22       0       0       0       0       0      0        0        0        0
#> 23       0       0       0       0       0      0        0        0        0
#> 24       0       0       0       0       0      0        0        0        0
#> 25       0       0       0       0       0      0        0        0        0
#> 26       0       0       0       0       0      0        0        0        0
#> 27       0       0       0       0       0      0        0        0        0
#> 28       0       0       0       0       0      0        0        0        0
#> 29       0       0       0       0       0      0        0        0        0
#> 30       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2      2     0     0     0     0
#> 3      2     1     0     0     0
#> 4      3     1     0     0     0
#> 5      4     1     0     0     0
#> 6      5     2     0     0     0
#> 7      5     2     0     0     0
#> 8      8     2     1     0     0
#> 9      8     2     1     0     0
#> 10     6     3     1     0     0
#> 11    NA    NA    NA    NA    NA
#> 12     1     0     0     0     0
#> 13     1     0     0     0     0
#> 14     4     0     0     0     0
#> 15     4     0     0     0     0
#> 16     3     1     0     0     0
#> 17     4     1     0     0     0
#> 18    10     1     0     1     0
#> 19    10     2     1     1     0
#> 20    12     2     1     1     0
#> 21    NA    NA    NA    NA    NA
#> 22     0     0     0     0     0
#> 23     0     0     1     0     0
#> 24     0     0     1     0     0
#> 25     0     0     1     0     0
#> 26     0     0     1     0     0
#> 27     0     0     1     0     0
#> 28     0     0     1     0     0
#> 29     0     0     1     0     0
#> 30     0     0     1     0     0

as.data.frame(mod, sim = 2)

#>    sim time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1    2    1   500     1 501       0       0       0         0       0       0
#> 2    2    2   499     1 501       1       0       0         0       0       0
#> 3    2    3   499     1 501       0       0       0         0       0       0
#> 4    2    4   496     1 501       1       0       0         0       0       0
#> 5    2    5   495     2 501       0       0       0         0       0       0
#> 6    2    6   494     3 501       1       1       0         0       0       0
#> 7    2    7   492     4 501       1       0       0         0       0       0
#> 8    2    8   485     4 501       4       0       0         1       0       0
#> 9    2    9   483     4 501       2       1       1         0       0       0
#> 10   2   10   478     7 501       2       0       0         0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2      1     0     0     0     0
#> 3      1     0     0     0     0
#> 4      4     0     0     0     0
#> 5      4     0     0     0     0
#> 6      3     1     0     0     0
#> 7      4     1     0     0     0
#> 8     10     1     0     1     0
#> 9     10     2     1     1     0
#> 10    12     2     1     1     0

# Time-specific means across simulations
as.data.frame(mod, out = "mean")

#>    time    s.num     i.num num   se.flow   is.flow   iq.flow iq2h.flow hf.flow
#> 1     1 500.0000 1.0000000 501 0.0000000 0.0000000 0.0000000 0.0000000       0
#> 2     2 499.0000 1.0000000 501 0.6666667 0.0000000 0.0000000 0.0000000       0
#> 3     3 498.6667 0.6666667 501 0.0000000 0.3333333 0.3333333 0.0000000       0
#> 4     4 497.3333 0.6666667 501 0.3333333 0.0000000 0.0000000 0.0000000       0
#> 5     5 496.3333 1.3333333 501 0.6666667 0.0000000 0.0000000 0.0000000       0
#> 6     6 495.0000 2.0000000 501 1.0000000 0.6666667 0.0000000 0.0000000       0
#> 7     7 494.0000 2.6666667 501 0.6666667 0.0000000 0.0000000 0.0000000       0
#> 8     8 490.3333 2.6666667 501 2.0000000 0.0000000 0.3333333 0.3333333       0
#> 9     9 489.0000 3.3333333 501 1.3333333 0.3333333 0.3333333 0.0000000       0
#> 10   10 487.0000 5.0000000 501 1.0000000 0.3333333 0.0000000 0.0000000       0
#>    ds.flow de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow
#> 1        0       0       0       0       0       0      0        0        0
#> 2        0       0       0       0       0       0      0        0        0
#> 3        0       0       0       0       0       0      0        0        0
#> 4        0       0       0       0       0       0      0        0        0
#> 5        0       0       0       0       0       0      0        0        0
#> 6        0       0       0       0       0       0      0        0        0
#> 7        0       0       0       0       0       0      0        0        0
#> 8        0       0       0       0       0       0      0        0        0
#> 9        0       0       0       0       0       0      0        0        0
#> 10       0       0       0       0       0       0      0        0        0
#>    a.q.flow    e.num     r.num     q.num     h.num f.num
#> 1         0      NaN       NaN       NaN       NaN   NaN
#> 2         0 1.000000 0.0000000 0.0000000 0.0000000     0
#> 3         0 1.000000 0.3333333 0.3333333 0.0000000     0
#> 4         0 2.333333 0.3333333 0.3333333 0.0000000     0
#> 5         0 2.666667 0.3333333 0.3333333 0.0000000     0
#> 6         0 2.666667 1.0000000 0.3333333 0.0000000     0
#> 7         0 3.000000 1.0000000 0.3333333 0.0000000     0
#> 8         0 6.000000 1.0000000 0.6666667 0.3333333     0
#> 9         0 6.000000 1.3333333 1.0000000 0.3333333     0
#> 10        0 6.000000 1.6666667 1.0000000 0.3333333     0

# Time-specific standard deviations across simulations
as.data.frame(mod, out = "sd")

#>    time     s.num     i.num num   se.flow   is.flow   iq.flow iq2h.flow hf.flow
#> 1     1  0.000000 0.0000000   0 0.0000000 0.0000000 0.0000000 0.0000000       0
#> 2     2  1.000000 0.0000000   0 0.5773503 0.0000000 0.0000000 0.0000000       0
#> 3     3  1.527525 0.5773503   0 0.0000000 0.5773503 0.5773503 0.0000000       0
#> 4     4  2.309401 0.5773503   0 0.5773503 0.0000000 0.0000000 0.0000000       0
#> 5     5  3.214550 1.1547005   0 1.1547005 0.0000000 0.0000000 0.0000000       0
#> 6     6  4.582576 1.7320508   0 1.0000000 0.5773503 0.0000000 0.0000000       0
#> 7     7  5.291503 2.3094011   0 0.5773503 0.0000000 0.0000000 0.0000000       0
#> 8     8  8.386497 2.3094011   0 2.0000000 0.0000000 0.5773503 0.5773503       0
#> 9     9  9.539392 3.0550505   0 1.1547005 0.5773503 0.5773503 0.0000000       0
#> 10   10 11.532563 4.3588989   0 1.0000000 0.5773503 0.0000000 0.0000000       0
#>    ds.flow de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow
#> 1        0       0       0       0       0       0      0        0        0
#> 2        0       0       0       0       0       0      0        0        0
#> 3        0       0       0       0       0       0      0        0        0
#> 4        0       0       0       0       0       0      0        0        0
#> 5        0       0       0       0       0       0      0        0        0
#> 6        0       0       0       0       0       0      0        0        0
#> 7        0       0       0       0       0       0      0        0        0
#> 8        0       0       0       0       0       0      0        0        0
#> 9        0       0       0       0       0       0      0        0        0
#> 10       0       0       0       0       0       0      0        0        0
#>    a.q.flow    e.num     r.num     q.num     h.num f.num
#> 1         0       NA        NA        NA        NA    NA
#> 2         0 1.000000 0.0000000 0.0000000 0.0000000     0
#> 3         0 1.000000 0.5773503 0.5773503 0.0000000     0
#> 4         0 2.081666 0.5773503 0.5773503 0.0000000     0
#> 5         0 2.309401 0.5773503 0.5773503 0.0000000     0
#> 6         0 2.516611 1.0000000 0.5773503 0.0000000     0
#> 7         0 2.645751 1.0000000 0.5773503 0.0000000     0
#> 8         0 5.291503 1.0000000 0.5773503 0.5773503     0
#> 9         0 5.291503 1.1547005 0.0000000 0.5773503     0
#> 10        0 6.000000 1.5275252 0.0000000 0.5773503     0

# Time-specific quantile values across simulations
as.data.frame(mod, out = "qnt", qval = 0.25)

#>    time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1     1 500.0   1.0 501     0.0     0.0       0         0       0       0
#> 2     2 498.5   1.0 501     0.5     0.0       0         0       0       0
#> 3     3 498.0   0.5 501     0.0     0.0       0         0       0       0
#> 4     4 496.0   0.5 501     0.0     0.0       0         0       0       0
#> 5     5 494.5   1.0 501     0.0     0.0       0         0       0       0
#> 6     6 492.5   1.5 501     0.5     0.5       0         0       0       0
#> 7     7 491.0   2.0 501     0.5     0.0       0         0       0       0
#> 8     8 485.5   2.0 501     1.0     0.0       0         0       0       0
#> 9     9 483.5   2.0 501     1.0     0.0       0         0       0       0
#> 10   10 480.5   3.5 501     0.5     0.0       0         0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2    0.5   0.0   0.0     0     0
#> 3    0.5   0.0   0.0     0     0
#> 4    1.5   0.0   0.0     0     0
#> 5    2.0   0.0   0.0     0     0
#> 6    1.5   0.5   0.0     0     0
#> 7    2.0   0.5   0.0     0     0
#> 8    4.0   0.5   0.5     0     0
#> 9    4.0   1.0   1.0     0     0
#> 10   3.0   1.0   1.0     0     0

as.data.frame(mod, out = "qnt", qval = 0.75)

#>    time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1     1 500.0   1.0 501     0.0     0.0     0.0       0.0       0       0
#> 2     2 499.5   1.0 501     1.0     0.0     0.0       0.0       0       0
#> 3     3 499.5   1.0 501     0.0     0.5     0.5       0.0       0       0
#> 4     4 498.0   1.0 501     0.5     0.0     0.0       0.0       0       0
#> 5     5 497.5   2.0 501     1.0     0.0     0.0       0.0       0       0
#> 6     6 497.0   3.0 501     1.5     1.0     0.0       0.0       0       0
#> 7     7 496.0   4.0 501     1.0     0.0     0.0       0.0       0       0
#> 8     8 493.0   4.0 501     3.0     0.0     0.5       0.5       0       0
#> 9     9 492.0   5.0 501     2.0     0.5     0.5       0.0       0       0
#> 10   10 491.5   7.5 501     1.5     0.5     0.0       0.0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2    1.5   0.0   0.0   0.0     0
#> 3    1.5   0.5   0.5   0.0     0
#> 4    3.5   0.5   0.5   0.0     0
#> 5    4.0   0.5   0.5   0.0     0
#> 6    4.0   1.5   0.5   0.0     0
#> 7    4.5   1.5   0.5   0.0     0
#> 8    9.0   1.5   1.0   0.5     0
#> 9    9.0   2.0   1.0   0.5     0
#> 10   9.0   2.5   1.0   0.5     0