Extract Model Data for Stochastic Models

This function extracts model simulations for objects of classes seiqhrf into a data frame using the generic as.data.frame function.

# S3 method for seiqhrf
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  out = "vals",
  sim,
  qval,
  ...
)

Arguments

x
An EpiModel object of class icm or netsim.

row.names
See as.data.frame.default.

optional
See as.data.frame.default.

out
Data output to data frame: "mean" for row means across simulations, "sd" for row standard deviations across simulations, "qnt" for row quantiles at the level specified in qval, or "vals" for values from individual simulations.

sim
If out="vals", the simulation number to output. If not specified, then data from all simulations will be output.

qval
Quantile value required when out="qnt".

...
See as.data.frame.default.

Details

These methods work for both seiqhrf class models. The available output includes time-specific means, standard deviations, quantiles, and simulation values (compartment and flow sizes) from these stochastic model classes. Means, standard deviations, and quantiles are calculated by taking the row summary (i.e., each row of data is corresponds to a time step) across all simulations in the model output.

Examples

## Stochastic SEIQHRF model
param <- param_seiqhrf()
init <- init_seiqhrf(s.num = 500, i.num = 1)
control <- control_seiqhrf(nsteps = 10, nsims = 3)
mod <- seiqhrf(init, control, param)

# Default output all simulation runs, default to all in stacked data.frame
as.data.frame(mod)

#>    sim time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1    1    1   500     1 501       0       0       0         0       0       0
#> 2    1    2   500     1 501       0       0       0         0       0       0
#> 3    1    3   498     1 501       2       0       0         0       0       0
#> 4    1    4   497     1 501       0       1       0         0       0       0
#> 5    1    5   496     1 501       0       0       0         0       0       0
#> 6    1    6   494     2 501       1       0       0         0       0       0
#> 7    1    7   493     3 501       0       0       0         0       0       0
#> 8    1    8   489     4 501       3       0       0         0       0       0
#> 9    1    9   485     5 501       2       1       0         0       0       0
#> 10   1   10   483     6 501       0       0       1         0       0       0
#> 11   2    1   500     1 501       0       0       0         0       0       0
#> 12   2    2   497     1 501       2       0       0         0       0       0
#> 13   2    3   496     1 501       1       0       0         0       0       0
#> 14   2    4   495     2 501       0       0       0         0       0       0
#> 15   2    5   492     3 501       1       3       0         0       0       0
#> 16   2    6   491     4 501       0       0       0         0       0       0
#> 17   2    7   491     5 501       0       0       0         0       0       0
#> 18   2    8   486     6 501       3       0       0         0       0       0
#> 19   2    9   483     5 501       2       2       1         1       0       0
#> 20   2   10   476     6 501       3       0       0         0       0       0
#> 21   3    1   500     1 501       0       0       0         0       0       0
#> 22   3    2   498     1 501       2       0       0         0       0       0
#> 23   3    3   497     1 501       1       0       0         0       0       0
#> 24   3    4   495     2 501       1       0       0         0       0       0
#> 25   3    5   491     3 501       3       0       0         0       0       0
#> 26   3    6   489     4 501       1       1       0         0       0       0
#> 27   3    7   479     6 501       6       0       0         0       0       0
#> 28   3    8   474     8 501       2       0       1         0       0       0
#> 29   3    9   468    11 501       4       0       1         0       0       0
#> 30   3   10   454    15 501       5       3       0         0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#> 11       0       0       0       0       0      0        0        0        0
#> 12       0       0       0       0       0      0        0        0        0
#> 13       0       0       0       0       0      0        0        0        0
#> 14       0       0       0       0       0      0        0        0        0
#> 15       0       0       0       0       0      0        0        0        0
#> 16       0       0       0       0       0      0        0        0        0
#> 17       0       0       0       0       0      0        0        0        0
#> 18       0       0       0       0       0      0        0        0        0
#> 19       0       0       0       0       0      0        0        0        0
#> 20       0       0       0       0       0      0        0        0        0
#> 21       0       0       0       0       0      0        0        0        0
#> 22       0       0       0       0       0      0        0        0        0
#> 23       0       0       0       0       0      0        0        0        0
#> 24       0       0       0       0       0      0        0        0        0
#> 25       0       0       0       0       0      0        0        0        0
#> 26       0       0       0       0       0      0        0        0        0
#> 27       0       0       0       0       0      0        0        0        0
#> 28       0       0       0       0       0      0        0        0        0
#> 29       0       0       0       0       0      0        0        0        0
#> 30       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2      0     0     0     0     0
#> 3      2     0     0     0     0
#> 4      2     1     0     0     0
#> 5      3     1     0     0     0
#> 6      4     1     0     0     0
#> 7      4     1     0     0     0
#> 8      7     1     0     0     0
#> 9      9     2     0     0     0
#> 10     9     2     1     0     0
#> 11    NA    NA    NA    NA    NA
#> 12     3     0     0     0     0
#> 13     4     0     0     0     0
#> 14     4     0     0     0     0
#> 15     3     3     0     0     0
#> 16     3     3     0     0     0
#> 17     2     3     0     0     0
#> 18     6     3     0     0     0
#> 19     6     5     1     1     0
#> 20    12     5     1     1     0
#> 21    NA    NA    NA    NA    NA
#> 22     2     0     0     0     0
#> 23     3     0     0     0     0
#> 24     4     0     0     0     0
#> 25     7     0     0     0     0
#> 26     7     1     0     0     0
#> 27    15     1     0     0     0
#> 28    17     1     1     0     0
#> 29    19     1     2     0     0
#> 30    26     4     2     0     0

as.data.frame(mod, sim = 2)

#>    sim time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1    2    1   500     1 501       0       0       0         0       0       0
#> 2    2    2   497     1 501       2       0       0         0       0       0
#> 3    2    3   496     1 501       1       0       0         0       0       0
#> 4    2    4   495     2 501       0       0       0         0       0       0
#> 5    2    5   492     3 501       1       3       0         0       0       0
#> 6    2    6   491     4 501       0       0       0         0       0       0
#> 7    2    7   491     5 501       0       0       0         0       0       0
#> 8    2    8   486     6 501       3       0       0         0       0       0
#> 9    2    9   483     5 501       2       2       1         1       0       0
#> 10   2   10   476     6 501       3       0       0         0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2      3     0     0     0     0
#> 3      4     0     0     0     0
#> 4      4     0     0     0     0
#> 5      3     3     0     0     0
#> 6      3     3     0     0     0
#> 7      2     3     0     0     0
#> 8      6     3     0     0     0
#> 9      6     5     1     1     0
#> 10    12     5     1     1     0

# Time-specific means across simulations
as.data.frame(mod, out = "mean")

#>    time    s.num    i.num num   se.flow   is.flow   iq.flow iq2h.flow hf.flow
#> 1     1 500.0000 1.000000 501 0.0000000 0.0000000 0.0000000 0.0000000       0
#> 2     2 498.3333 1.000000 501 1.3333333 0.0000000 0.0000000 0.0000000       0
#> 3     3 497.0000 1.000000 501 1.3333333 0.0000000 0.0000000 0.0000000       0
#> 4     4 495.6667 1.666667 501 0.3333333 0.3333333 0.0000000 0.0000000       0
#> 5     5 493.0000 2.333333 501 1.3333333 1.0000000 0.0000000 0.0000000       0
#> 6     6 491.3333 3.333333 501 0.6666667 0.3333333 0.0000000 0.0000000       0
#> 7     7 487.6667 4.666667 501 2.0000000 0.0000000 0.0000000 0.0000000       0
#> 8     8 483.0000 6.000000 501 2.6666667 0.0000000 0.3333333 0.0000000       0
#> 9     9 478.6667 7.000000 501 2.6666667 1.0000000 0.6666667 0.3333333       0
#> 10   10 471.0000 9.000000 501 2.6666667 1.0000000 0.3333333 0.0000000       0
#>    ds.flow de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow
#> 1        0       0       0       0       0       0      0        0        0
#> 2        0       0       0       0       0       0      0        0        0
#> 3        0       0       0       0       0       0      0        0        0
#> 4        0       0       0       0       0       0      0        0        0
#> 5        0       0       0       0       0       0      0        0        0
#> 6        0       0       0       0       0       0      0        0        0
#> 7        0       0       0       0       0       0      0        0        0
#> 8        0       0       0       0       0       0      0        0        0
#> 9        0       0       0       0       0       0      0        0        0
#> 10       0       0       0       0       0       0      0        0        0
#>    a.q.flow     e.num     r.num     q.num     h.num f.num
#> 1         0       NaN       NaN       NaN       NaN   NaN
#> 2         0  1.666667 0.0000000 0.0000000 0.0000000     0
#> 3         0  3.000000 0.0000000 0.0000000 0.0000000     0
#> 4         0  3.333333 0.3333333 0.0000000 0.0000000     0
#> 5         0  4.333333 1.3333333 0.0000000 0.0000000     0
#> 6         0  4.666667 1.6666667 0.0000000 0.0000000     0
#> 7         0  7.000000 1.6666667 0.0000000 0.0000000     0
#> 8         0 10.000000 1.6666667 0.3333333 0.0000000     0
#> 9         0 11.333333 2.6666667 1.0000000 0.3333333     0
#> 10        0 15.666667 3.6666667 1.3333333 0.3333333     0

# Time-specific standard deviations across simulations
as.data.frame(mod, out = "sd")

#>    time     s.num     i.num num   se.flow   is.flow   iq.flow iq2h.flow hf.flow
#> 1     1  0.000000 0.0000000   0 0.0000000 0.0000000 0.0000000 0.0000000       0
#> 2     2  1.527525 0.0000000   0 1.1547005 0.0000000 0.0000000 0.0000000       0
#> 3     3  1.000000 0.0000000   0 0.5773503 0.0000000 0.0000000 0.0000000       0
#> 4     4  1.154701 0.5773503   0 0.5773503 0.5773503 0.0000000 0.0000000       0
#> 5     5  2.645751 1.1547005   0 1.5275252 1.7320508 0.0000000 0.0000000       0
#> 6     6  2.516611 1.1547005   0 0.5773503 0.5773503 0.0000000 0.0000000       0
#> 7     7  7.571878 1.5275252   0 3.4641016 0.0000000 0.0000000 0.0000000       0
#> 8     8  7.937254 2.0000000   0 0.5773503 0.0000000 0.5773503 0.0000000       0
#> 9     9  9.291573 3.4641016   0 1.1547005 1.0000000 0.5773503 0.5773503       0
#> 10   10 15.132746 5.1961524   0 2.5166115 1.7320508 0.5773503 0.0000000       0
#>    ds.flow de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow
#> 1        0       0       0       0       0       0      0        0        0
#> 2        0       0       0       0       0       0      0        0        0
#> 3        0       0       0       0       0       0      0        0        0
#> 4        0       0       0       0       0       0      0        0        0
#> 5        0       0       0       0       0       0      0        0        0
#> 6        0       0       0       0       0       0      0        0        0
#> 7        0       0       0       0       0       0      0        0        0
#> 8        0       0       0       0       0       0      0        0        0
#> 9        0       0       0       0       0       0      0        0        0
#> 10       0       0       0       0       0       0      0        0        0
#>    a.q.flow    e.num     r.num     q.num     h.num f.num
#> 1         0       NA        NA        NA        NA    NA
#> 2         0 1.527525 0.0000000 0.0000000 0.0000000     0
#> 3         0 1.000000 0.0000000 0.0000000 0.0000000     0
#> 4         0 1.154701 0.5773503 0.0000000 0.0000000     0
#> 5         0 2.309401 1.5275252 0.0000000 0.0000000     0
#> 6         0 2.081666 1.1547005 0.0000000 0.0000000     0
#> 7         0 7.000000 1.1547005 0.0000000 0.0000000     0
#> 8         0 6.082763 1.1547005 0.5773503 0.0000000     0
#> 9         0 6.806859 2.0816660 1.0000000 0.5773503     0
#> 10        0 9.073772 1.5275252 0.5773503 0.5773503     0

# Time-specific quantile values across simulations
as.data.frame(mod, out = "qnt", qval = 0.25)

#>    time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1     1 500.0   1.0 501     0.0     0.0     0.0         0       0       0
#> 2     2 497.5   1.0 501     1.0     0.0     0.0         0       0       0
#> 3     3 496.5   1.0 501     1.0     0.0     0.0         0       0       0
#> 4     4 495.0   1.5 501     0.0     0.0     0.0         0       0       0
#> 5     5 491.5   2.0 501     0.5     0.0     0.0         0       0       0
#> 6     6 490.0   3.0 501     0.5     0.0     0.0         0       0       0
#> 7     7 485.0   4.0 501     0.0     0.0     0.0         0       0       0
#> 8     8 480.0   5.0 501     2.5     0.0     0.0         0       0       0
#> 9     9 475.5   5.0 501     2.0     0.5     0.5         0       0       0
#> 10   10 465.0   6.0 501     1.5     0.0     0.0         0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2    1.0   0.0   0.0     0     0
#> 3    2.5   0.0   0.0     0     0
#> 4    3.0   0.0   0.0     0     0
#> 5    3.0   0.5   0.0     0     0
#> 6    3.5   1.0   0.0     0     0
#> 7    3.0   1.0   0.0     0     0
#> 8    6.5   1.0   0.0     0     0
#> 9    7.5   1.5   0.5     0     0
#> 10  10.5   3.0   1.0     0     0

as.data.frame(mod, out = "qnt", qval = 0.75)

#>    time s.num i.num num se.flow is.flow iq.flow iq2h.flow hf.flow ds.flow
#> 1     1 500.0   1.0 501     0.0     0.0     0.0       0.0       0       0
#> 2     2 499.0   1.0 501     2.0     0.0     0.0       0.0       0       0
#> 3     3 497.5   1.0 501     1.5     0.0     0.0       0.0       0       0
#> 4     4 496.0   2.0 501     0.5     0.5     0.0       0.0       0       0
#> 5     5 494.0   3.0 501     2.0     1.5     0.0       0.0       0       0
#> 6     6 492.5   4.0 501     1.0     0.5     0.0       0.0       0       0
#> 7     7 492.0   5.5 501     3.0     0.0     0.0       0.0       0       0
#> 8     8 487.5   7.0 501     3.0     0.0     0.5       0.0       0       0
#> 9     9 484.0   8.0 501     3.0     1.5     1.0       0.5       0       0
#> 10   10 479.5  10.5 501     4.0     1.5     0.5       0.0       0       0
#>    de.flow di.flow dq.flow dh.flow dr.flow a.flow a.e.flow a.i.flow a.q.flow
#> 1        0       0       0       0       0      0        0        0        0
#> 2        0       0       0       0       0      0        0        0        0
#> 3        0       0       0       0       0      0        0        0        0
#> 4        0       0       0       0       0      0        0        0        0
#> 5        0       0       0       0       0      0        0        0        0
#> 6        0       0       0       0       0      0        0        0        0
#> 7        0       0       0       0       0      0        0        0        0
#> 8        0       0       0       0       0      0        0        0        0
#> 9        0       0       0       0       0      0        0        0        0
#> 10       0       0       0       0       0      0        0        0        0
#>    e.num r.num q.num h.num f.num
#> 1     NA    NA    NA    NA    NA
#> 2    2.5   0.0   0.0   0.0     0
#> 3    3.5   0.0   0.0   0.0     0
#> 4    4.0   0.5   0.0   0.0     0
#> 5    5.0   2.0   0.0   0.0     0
#> 6    5.5   2.0   0.0   0.0     0
#> 7    9.5   2.0   0.0   0.0     0
#> 8   12.0   2.0   0.5   0.0     0
#> 9   14.0   3.5   1.5   0.5     0
#> 10  19.0   4.5   1.5   0.5     0